# api/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Video Platform API",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000/v1"
        }
    ],
    "paths": {
...
        "/auth/email/verify": {
            "post": {
                "operationId": "auth.verifyEmail",
                "tags": [
                    "Auth"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "token": {
                                        "type": "string",
                                        "maxLength": 128
                                    }
                                },
                                "required": [
                                    "token"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "email_verified": {
                                            "type": "boolean"
                                        }
                                    },
                                    "required": [
                                        "email_verified"
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "type": {
                                            "type": "string",
                                            "description": "URI identifying the problem type."
                                        },
                                        "title": {
                                            "type": "string"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "examples": [
                                                422
                                            ]
                                        },
                                        "detail": {
                                            "type": "string"
                                        },
                                        "code": {
                                            "type": "string",
                                            "description": "Stable machine-readable error code."
                                        },
                                        "request_id": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "field": {
                                                        "type": "string"
                                                    },
                                                    "code": {
                                                        "type": "string"
                                                    },
                                                    "message": {
                                                        "type": "string"
                                                    }
                                                },
                                                "required": [
                                                    "field",
                                                    "code",
                                                    "message"
                                                ]
                                            }
                                        }
                                    },
                                    "required": [
                                        "type",
                                        "title",
                                        "status",
                                        "code"
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/email/resend": {
            "post": {
                "operationId": "auth.resendVerification",
                "summary": "Always 202, whatever the email's state",
                "tags": [
                    "Auth"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "maxLength": 254
                                    }
                                },
                                "required": [
                                    "email"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "const": "If this email needs verifying, we sent a new link."
                                        }
                                    },
                                    "required": [
                                        "message"
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "type": {
                                            "type": "string",
                                            "description": "URI identifying the problem type."
                                        },
                                        "title": {
                                            "type": "string"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "examples": [
                                                422
                                            ]
                                        },
                                        "detail": {
                                            "type": "string"
                                        },
                                        "code": {
                                            "type": "string",
                                            "description": "Stable machine-readable error code."
                                        },
                                        "request_id": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "field": {
                                                        "type": "string"
                                                    },
                                                    "code": {
                                                        "type": "string"
                                                    },
                                                    "message": {
                                                        "type": "string"
                                                    }
                                                },
                                                "required": [
                                                    "field",
                                                    "code",
                                                    "message"
                                                ]
                                            }
                                        }
                                    },
                                    "required": [
                                        "type",
                                        "title",
                                        "status",
                                        "code"
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/auth/logout-all": {
            "post": {
                "operationId": "auth.logoutAll",
                "tags": [
                    "Auth"
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "type": {
                                            "type": "string",
                                            "description": "URI identifying the problem type."
                                        },
                                        "title": {
                                            "type": "string"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "examples": [
                                                401
                                            ]
                                        },
                                        "detail": {
                                            "type": "string"
                                        },
                                        "code": {
                                            "type": "string",
                                            "description": "Stable machine-readable error code."
                                        },
                                        "request_id": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "field": {
                                                        "type": "string"
                                                    },
                                                    "code": {
                                                        "type": "string"
                                                    },
                                                    "message": {
                                                        "type": "string"
                                                    }
                                                },
                                                "required": [
                                                    "field",
                                                    "code",
                                                    "message"
                                                ]
                                            }
                                        }
                                    },
                                    "required": [
                                        "type",
                                        "title",
                                        "status",
                                        "code"
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}